rename repo and packages
