+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": null,
     "lines": null
   },
-  "generated_at": "2019-12-03T21:42:45Z",
+  "generated_at": "2019-12-11T14:13:30Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
