+++ datapump_utils/dataset.py
@@ -1,9 +1,9 @@
 import requests
 import json
 
-from geotrellis_summary_update.secrets import get_token
-from geotrellis_summary_update.util import api_prefix
-from geotrellis_summary_update.exceptions import UnexpectedResponseError
+from datapump_utils.secrets import get_token
+from datapump_utils.util import api_prefix
+from datapump_utils.exceptions import UnexpectedResponseError
 
 
 def get_dataset(dataset_id):
