+++ datapump_utils/summary_analysis.py
@@ -4,8 +4,8 @@ from enum import Enum
 import boto3
 from botocore.exceptions import ClientError
 
-from geotrellis_summary_update.util import bucket_suffix
-from geotrellis_summary_update.s3 import get_s3_path, s3_client
+from datapump_utils.util import bucket_suffix
+from datapump_utils.s3 import get_s3_path, s3_client
 
 
 RESULT_BUCKET = f"gfw-pipelines{bucket_suffix()}"
