+++ terraform/data.tf
@@ -7,11 +7,11 @@ data "template_file" "sts_assume_role_states" {
 }
 
 data "template_file" "geotrellis_summary_update" {
-  template = file("policies/geotrellis_summary_update.json")
+  template = file("policies/datapump.json")
 }
 
 data "template_file" "sfn_geotrellis_summary_update" {
-  template = file("../step_functions/geotrellis_summary_update.json")
+  template = file("../step_functions/geotrellis_dataset.json")
   vars = {
     lambda_submit_job_arn     = aws_lambda_function.submit_job.arn,
     lambda_upload_results_arn = aws_lambda_function.upload_results_to_datasets.arn,
@@ -19,12 +19,12 @@ data "template_file" "sfn_geotrellis_summary_update" {
   }
 }
 
-data "template_file" "sfn_summarize_new_areas" {
-  template = file("../step_functions/summarize_new_areas.json")
+data "template_file" "sfn_summarize_new_aoi" {
+  template = file("../step_functions/new_user_aoi.json")
   vars = {
-    lambda_check_new_areas_arn        = aws_lambda_function.check_new_areas.arn,
-    lambda_update_new_area_status_arn = aws_lambda_function.update_new_area_statuses.arn,
-    state_machine_arn                 = aws_sfn_state_machine.geotrellis_summary_update.id
+    lambda_check_new_aoi_arn        = aws_lambda_function.check_new_aoi.arn,
+    lambda_update_new_aoi_status_arn = aws_lambda_function.update_new_aoi_statuses.arn,
+    state_machine_arn                 = aws_sfn_state_machine.geotrellis_dataset.id
   }
 }
