+++ terraform/lambdas.tf
@@ -1,7 +1,7 @@
-resource "aws_lambda_layer_version" "geotrellis_summary_update" {
-  layer_name          = "geotrellis_summary_update"
-  filename            = local.lambda_layer_geotrellis_summary_update
-  source_code_hash    = filebase64sha256(local.lambda_layer_geotrellis_summary_update)
+resource "aws_lambda_layer_version" "datapump_utils" {
+  layer_name          = "datapump_utils"
+  filename            = local.lambda_layer_datapump_utils
+  source_code_hash    = filebase64sha256(local.lambda_layer_datapump_utils)
   compatible_runtimes = [var.lambda_submit_job_runtime]
 }
 
@@ -13,17 +13,17 @@ resource "aws_lambda_layer_version" "shapely_pyyaml" {
 }
 
 resource "aws_lambda_function" "submit_job" {
-  function_name    = "submit-job_geotrellis-summary-update${local.name_suffix}"
+  function_name    = "${local.project}-submit_job${local.name_suffix}"
   filename         = data.archive_file.lambda_submit_job.output_path
   source_code_hash = data.archive_file.lambda_submit_job.output_base64sha256
-  role             = aws_iam_role.geotrellis_summary_update_lambda.arn
+  role             = aws_iam_role.datapump_lambda.arn
   runtime          = var.lambda_submit_job_runtime
   handler          = "lambda_function.handler"
   memory_size      = var.lambda_submit_job_memory_size
   timeout          = var.lambda_submit_job_timeout
   publish          = true
   tags             = local.tags
-  layers           = [aws_lambda_layer_version.geotrellis_summary_update.arn]
+  layers           = [aws_lambda_layer_version.datapump_utils.arn]
   environment {
     variables = {
       ENV                = var.environment
@@ -34,17 +34,17 @@ resource "aws_lambda_function" "submit_job" {
 }
 
 resource "aws_lambda_function" "upload_results_to_datasets" {
-  function_name    = "upload-results-to-datasets_geotrellis-summary-update${local.name_suffix}"
+  function_name    = "${local.project}-upload_results_to_datasets${local.name_suffix}"
   filename         = data.archive_file.lambda_upload_results_to_datasets.output_path
   source_code_hash = data.archive_file.lambda_upload_results_to_datasets.output_base64sha256
-  role             = aws_iam_role.geotrellis_summary_update_lambda.arn
+  role             = aws_iam_role.datapump_lambda.arn
   runtime          = var.lambda_upload_results_runtime
   handler          = "lambda_function.handler"
   memory_size      = var.lambda_upload_results_memory_size
   timeout          = var.lambda_upload_results_timeout
   publish          = true
   tags             = local.tags
-  layers           = [aws_lambda_layer_version.geotrellis_summary_update.arn]
+  layers           = [aws_lambda_layer_version.datapump_utils.arn]
   environment {
     variables = {
       ENV = var.environment
@@ -53,17 +53,17 @@ resource "aws_lambda_function" "upload_results_to_datasets" {
 }
 
 resource "aws_lambda_function" "check_datasets_saved" {
-  function_name    = "check-datasets-saved_geotrellis-summary-update${local.name_suffix}"
+  function_name    = "${local.project}-check_datasets_saved${local.name_suffix}"
   filename         = data.archive_file.lambda_check_datasets_saved.output_path
   source_code_hash = data.archive_file.lambda_check_datasets_saved.output_base64sha256
-  role             = aws_iam_role.geotrellis_summary_update_lambda.arn
+  role             = aws_iam_role.datapump_lambda.arn
   runtime          = var.lambda_check_datasets_runtime
   handler          = "lambda_function.handler"
   memory_size      = var.lambda_check_datasets_memory_size
   timeout          = var.lambda_check_datasets_timeout
   publish          = true
   tags             = local.tags
-  layers           = [aws_lambda_layer_version.geotrellis_summary_update.arn]
+  layers           = [aws_lambda_layer_version.datapump_utils.arn]
   environment {
     variables = {
       ENV = var.environment
@@ -71,18 +71,18 @@ resource "aws_lambda_function" "check_datasets_saved" {
   }
 }
 
-resource "aws_lambda_function" "check_new_areas" {
-  function_name    = "check-new-areas_geotrellis-summary-update${local.name_suffix}"
-  filename         = data.archive_file.lambda_check_new_areas.output_path
-  source_code_hash = data.archive_file.lambda_check_new_areas.output_base64sha256
-  role             = aws_iam_role.geotrellis_summary_update_lambda.arn
-  runtime          = var.lambda_check_new_areas_runtime
+resource "aws_lambda_function" "check_new_aoi" {
+  function_name    = "${local.project}-check_new_aoi${local.name_suffix}"
+  filename         = data.archive_file.lambda_check_new_aoi.output_path
+  source_code_hash = data.archive_file.lambda_check_new_aoi.output_base64sha256
+  role             = aws_iam_role.datapump_lambda.arn
+  runtime          = var.lambda_check_new_aoi_runtime
   handler          = "lambda_function.handler"
-  memory_size      = var.lambda_check_new_areas_memory_size
-  timeout          = var.lambda_check_new_areas_timeout
+  memory_size      = var.lambda_check_new_aoi_memory_size
+  timeout          = var.lambda_check_new_aoi_timeout
   publish          = true
   tags             = local.tags
-  layers           = [aws_lambda_layer_version.geotrellis_summary_update.arn, aws_lambda_layer_version.shapely_pyyaml.arn]
+  layers           = [aws_lambda_layer_version.datapump_utils.arn, aws_lambda_layer_version.shapely_pyyaml.arn]
   environment {
     variables = {
       ENV                = var.environment
@@ -92,18 +92,18 @@ resource "aws_lambda_function" "check_new_areas" {
   }
 }
 
-resource "aws_lambda_function" "update_new_area_statuses" {
-  function_name    = "update-new-area-statuses_geotrellis-summary-update${local.name_suffix}"
-  filename         = data.archive_file.lambda_update_new_area_statuses.output_path
-  source_code_hash = data.archive_file.lambda_update_new_area_statuses.output_base64sha256
-  role             = aws_iam_role.geotrellis_summary_update_lambda.arn
-  runtime          = var.lambda_update_new_area_statuses_runtime
+resource "aws_lambda_function" "update_new_aoi_statuses" {
+  function_name    = "${local.project}-update_new_aoi_statuses${local.name_suffix}"
+  filename         = data.archive_file.lambda_update_new_aoi_statuses.output_path
+  source_code_hash = data.archive_file.lambda_update_new_aoi_statuses.output_base64sha256
+  role             = aws_iam_role.datapump_lambda.arn
+  runtime          = var.lambda_update_new_aoi_statuses_runtime
   handler          = "lambda_function.handler"
-  memory_size      = var.lambda_update_new_area_statuses_memory_size
-  timeout          = var.lambda_update_new_area_statuses_timeout
+  memory_size      = var.lambda_update_new_aoi_statuses_memory_size
+  timeout          = var.lambda_update_new_aoi_statuses_timeout
   publish          = true
   tags             = local.tags
-  layers           = [aws_lambda_layer_version.geotrellis_summary_update.arn]
+  layers           = [aws_lambda_layer_version.datapump_utils.arn]
   environment {
     variables = {
       ENV                = var.environment
