+++ terraform/outputs.tf
@@ -1,25 +1,25 @@
 output "sfn_geotrellis_summary_update" {
-  value = aws_sfn_state_machine.geotrellis_summary_update.id
+  value = aws_sfn_state_machine.geotrellis_dataset.id
 }
 
-output "sfn_summarize_new_areas" {
-  value = aws_sfn_state_machine.summarize_new_areas.id
+output "sfn_summarize_new_aoi" {
+  value = aws_sfn_state_machine.new_user_aoi.id
 }
 
 output "lambda_check_new_dataset_daved" {
   value = aws_lambda_function.check_datasets_saved.id
 }
 
-output "lambda_check_new_area" {
-  value = aws_lambda_function.check_new_areas.id
+output "lambda_check_new_aoi" {
+  value = aws_lambda_function.check_new_aoi.id
 }
 
 output "lambda_submit_job" {
   value = aws_lambda_function.submit_job.id
 }
 
-output "lambda_update_new_area_statuses" {
-  value = aws_lambda_function.update_new_area_statuses.id
+output "lambda_update_new_aoi_statuses" {
+  value = aws_lambda_function.update_new_aoi_statuses.id
 }
 
 output "lambda_upload_results_to_datasets" {
