+++ tests/lambdas/test_check_new_areas.py
@@ -1,8 +1,8 @@
 import os
 
-from geotrellis_summary_update.util import secret_suffix
+from datapump_utils.util import secret_suffix
 
-from lambdas.check_new_areas.src.lambda_function import (
+from lambdas.check_new_aoi.src.lambda_function import (
     geostore_to_wkb,
     get_geostore,
     get_geostore_ids,
